Add in update getgter for messageType, chatId and from

+++ src/main/scala/PPS19/scalagram/models/ChatId.scala
@@ -1,10 +1,10 @@
 package PPS19.scalagram.models
 
-class ChatId private (chatId: Either[String, Int]) {
+class ChatId private (chatId: Either[String, Long]) {
   def get: Any = chatId.fold(l => l, r => r)
 }
 
 object ChatId {
   def apply(chatId: String) = new ChatId(Left(chatId))
-  def apply(chatId: Int) = new ChatId(Right(chatId))
+  def apply(chatId: Long) = new ChatId(Right(chatId))
 }
